delete user requires authentication in request

+++ api/controllers/UserController.php
@@ -40,34 +40,18 @@ class UserController
         return $response;
     }
 
-    // public function createUser(Array $input)
-    // {
-    //     $result = $this->userGateway->insert($input);
-    //     if(!$result)
-    //     {
-    //         $response = internalServerErrorResponse();
-    //     }
-    //     $response['status_code_header'] = 'HTTP/1.1 201 Created';
-    //     $response['body'] = json_encode(array('success' => true));
-    //     return $response;
-    // }
-
-    public function deleteUserById($id)
+    public function deleteUser($id, $password)
     {
+        $result = $this->userGateway->getPassword($id);
+        $result = pg_fetch_assoc($result);
+        $hashed_password = $result['pass'];
+        if(!password_verify($password, $hashed_password))
+        {
+            $response = unauthorizedResponse('Could not authenticate.');
+            return $response;
+        }
+        
         $result = $this->userGateway->delete($id);
-        $response = $this->handleDeleteUserResponse($result);
-        return $response;
-    }
-
-    public function deleteUserByUsername($username)
-    {
-        $result = $this->userGateway->delete(0, $username);
-        $response = $this->handleDeleteUserResponse($result);
-        return $response;
-    }
-
-    private function handleDeleteUserResponse($result)
-    {
         if(!$result)
         {
             $response = internalServerErrorResponse();
